Fixing incorrect uniqueIdentifier generation and organizing existing code
