+++ components/apimgt/org.wso2.carbon.apimgt.gateway/src/main/java/org/wso2/carbon/apimgt/gateway/mediators/oauth/OAuthTokenGenerator.java
@@ -20,14 +20,12 @@ package org.wso2.carbon.apimgt.gateway.mediators.oauth;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.json.simple.JSONObject;
 import org.wso2.carbon.apimgt.api.APIManagementException;
 import org.wso2.carbon.apimgt.gateway.handlers.security.APISecurityConstants;
 import org.wso2.carbon.apimgt.gateway.handlers.security.APISecurityException;
 import org.wso2.carbon.apimgt.gateway.mediators.oauth.client.OAuthClient;
 import org.wso2.carbon.apimgt.gateway.mediators.oauth.client.TokenResponse;
 import org.wso2.carbon.apimgt.gateway.mediators.oauth.conf.OAuthEndpoint;
-import org.wso2.carbon.apimgt.impl.APIConstants;
 
 import java.io.IOException;
 import java.util.concurrent.CountDownLatch;
@@ -35,12 +33,13 @@ import java.util.concurrent.CountDownLatch;
 public class OAuthTokenGenerator {
     private static final Log log = LogFactory.getLog(OAuthTokenGenerator.class);
 
-    public void checkTokenValidity(OAuthEndpoint oAuthEndpoint, CountDownLatch latch, boolean isRedisEnabled, RedisCache redisCache)
+    public void checkTokenValidity(OAuthEndpoint oAuthEndpoint, CountDownLatch latch, boolean isRedisEnabled,
+            RedisTokenCache redisTokenCache)
             throws APISecurityException {
         try {
             TokenResponse previousResponse = null;
             if (isRedisEnabled) {
-                previousResponse = redisCache.getTokenResponseById(oAuthEndpoint.getId());
+                previousResponse = redisTokenCache.getTokenResponseById(oAuthEndpoint.getId());
             } else {
                 previousResponse = TokenCache.getInstance().getTokenMap().get(oAuthEndpoint.getId());
             }
@@ -51,13 +50,14 @@ public class OAuthTokenGenerator {
 
                 if (timeDifference <= 1) {
                     if (previousResponse.getRefreshToken() != null) {
-                        addTokenToCache(oAuthEndpoint, previousResponse.getRefreshToken(), isRedisEnabled, redisCache);
+                        addTokenToCache(oAuthEndpoint, previousResponse.getRefreshToken(), isRedisEnabled,
+                                redisTokenCache);
                     } else {
-                        addTokenToCache(oAuthEndpoint, null, isRedisEnabled, redisCache);
+                        addTokenToCache(oAuthEndpoint, null, isRedisEnabled, redisTokenCache);
                     }
                 }
             } else {
-                addTokenToCache(oAuthEndpoint, null, isRedisEnabled, redisCache);
+                addTokenToCache(oAuthEndpoint, null, isRedisEnabled, redisTokenCache);
             }
         } catch (IOException e) {
             log.error("Error while generating OAuth Token" + getEndpointId(oAuthEndpoint));
@@ -71,7 +71,8 @@ public class OAuthTokenGenerator {
         latch.countDown();
     }
 
-    private void addTokenToCache(OAuthEndpoint oAuthEndpoint, String refreshToken, boolean isRedisEnabled, RedisCache redisCache)
+    private void addTokenToCache(OAuthEndpoint oAuthEndpoint, String refreshToken, boolean isRedisEnabled,
+            RedisTokenCache redisTokenCache)
             throws IOException, APIManagementException {
         TokenResponse tokenResponse = OAuthClient.generateToken(oAuthEndpoint.getTokenApiUrl(),
                 oAuthEndpoint.getClientId(), oAuthEndpoint.getClientSecret(), oAuthEndpoint.getUsername(),
@@ -79,7 +80,7 @@ public class OAuthTokenGenerator {
                 refreshToken);
         assert tokenResponse != null;
         if (isRedisEnabled) {
-            redisCache.addTokenResponse(oAuthEndpoint.getId(), tokenResponse);
+            redisTokenCache.addTokenResponse(oAuthEndpoint.getId(), tokenResponse);
         } else {
             TokenCache.getInstance().getTokenMap().put(oAuthEndpoint.getId(), tokenResponse);
         }
